// app/0/SiteHeaderWh 2/EditHeaderMenu/type.ts
import { sidebarTextInput , colorPalette} from "@components/sidebar/components/settings";

export interface EditHeaderMenuType {
  bgColor: colorPalette;

  title1: sidebarTextInput;
  link1: sidebarTextInput;

  title2: sidebarTextInput;
  link2: sidebarTextInput;

  title3: sidebarTextInput;
  link3: sidebarTextInput;

  title4: sidebarTextInput;
  link4: sidebarTextInput;

  title5: sidebarTextInput;
  link5: sidebarTextInput;
}